how do I config?
